% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{sidrs-package}
\alias{sidrs}
\alias{sidrs-package}
\title{sidrs: Compute the Structural Intervention Distance (SID) for DAGs.}
\description{
Compute the structural intervention distance for DAGs.
}
\author{
\strong{Maintainer}: Alessio Zanga \email{alessio.zanga@outlook.it} (\href{https://orcid.org/0000-0003-4423-2121}{ORCID})

}
